refactor(jina): 폼 ui를  feature 디렉토리로 이동 #61
