+++ src/features/createPost/ui/CreateForm/ImageInput.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { useState } from 'react';
 
-import { useImageInput } from 'features/community/createPost/model/useImageInput';
+import { useImageInput } from 'features/createPost/model/useImageInput';
 
 import { Button, Modal } from 'shared/ui/';
